+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+### Added
+- Added permissions for authorizing with apps
+
 ## 2.1.1 - 2019-10-08
 ### Added
 - Added `AuthorizationUrlEvent` event type